3.2: score and scores now in reducer
